requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
attrs==24.3.0
audioread==3.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
decouple==0.0.7
deepdiff==8.1.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.6
filelock==3.13.1
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.2.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-anthropic==0.3.1
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.6
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.23.2
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.58.1
openpyxl==3.1.5
orderly-set==5.2.3
orjson==3.10.13
packaging==24.2
pathlib==1.0.1
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydub==0.25.1
pyngrok==7.2.2
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websockets==14.1
XlsxWriter==3.2.0
yarl==1.18.3
